Fixed merge errors

+++ src/utils/config.py
@@ -164,7 +164,7 @@ class ConfigArguments:
         self.max_dimension = self.dimension
         if (self.record_length_resize>0):
             self.max_dimension =  int(math.sqrt(self.record_length_resize / 8))
-        self.resized_image = np.random.random((self.max_dimension, self.max_dimension))
+        self.resized_image = np.random.random((self.max_dimension, self.max_dimension)).astype('uint8')
         self.total_samples_train = self.num_samples_per_file * len(self.file_list_train)
         self.total_samples_eval = self.num_samples_per_file * len(self.file_list_eval)
         self.required_samples = self.comm_size * self.batch_size
